dvc: automatically create scripts, v2
Moving dvc.py contents to dvc/main.py and making
dvc.py to be just a wrapper for calling dvc/main.py.
This will allow us to both keep changes minimal and
use automatic script building.
